Moved logic to retrieve products info to ProductsInfoController
